+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   The error when you do not add an `icon_path` or have this option targeting a non existing file will now be more explicit, rather than being cryptic (because thrown by a dependency).
+
 ## [0.3.1] 2019-12-28
 
 ### Fixed